ENT-5684 Reconnect the progress tracker when unpausing the flow
Previously we were just throwing this away when pausing, meaning
updates would not be passed back to the user.
